Test the push on Travis
Now Travis as well as production is almost the same code path, the only
difference is that on Travis we push into a test repository, that is otherwise
not used.

We currently have some if statements for Travis in the docker container (in
scheduler.py as well as in update.sh), which later we should remove and
parametrize everything, i.e. pass the push repository as an argument, pass the
frequency of updates as well as count (e.g. 2x or infinity) as arguments. Then
on Travis those arguments do "docker run" are set differently than in
production, but that would be the only difference, the docker container itself
would always execute using the same code path.