Add name to the blog title
Fixes https://github.com/sympy/planet-sympy/issues/36

I haven't been able to test this, so please make sure it works. I think an
alternate way to fix this would be to edit the rss script (in case you don't
want the name to appear this way on the planet page itself).